Added Boilerplate DAO classes
Added all of the boilerplate DAO stuff. Should basically be ready for front-end stuff. We might also want to set a package up for utility classes that interact with the DAO classes that the front end can call.

As these are boilerplate functions, I wasn't entirely paying attention, and I haven't tested them, so be wary.

Also, small issue in the SetupSQL. I accidently had added the dateTaken to the primary key of testTaken lol. Fixed.